+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.0.5 / Unreleased
 
+* Makes game playable when logged in only
 * Fixes border rendering for small maps
 * Prohibits newly created units to move
 * Renders maps the same as in the map editor